Stops build early if the change only effects documentation (#40)
Prevent test and SNAPSHOT deploy of a documentation-only change.

if: conditions are preferred as they obviate job creation. However, they can only use
pre-defined env variables or globals that can be evaluated statically. The other way to skip
is to exit early. We do that here to prevent overhead of running tests and SNAPSHOT deploys
that only include Markdown (documentation) updates. Early terminating here will also prevent
the subsequent deploy SNAPSHOT job.

Note: This early exit does not prevent a release tag on a doc-only commit because the if:
conditions here don't match the release conditions.